requirements.txt
google-adk
google-cloud-aiplatform
google-cloud-texttospeech>=2.29.0
google-generativeai